refactor: sort 폴더 하위 불필요한 폴더 제거
